Port pos_payment_terminal to v10
