+++ pos_payment_terminal/demo/pos_payment_terminal_demo.xml
