+++ pos_payment_terminal/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from . import pos_config
+from . import account_journal
