--- pos_payment_terminal/pos_payment_terminal_view.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="view_pos_config_form" model="ir.ui.view">
-            <field name="name">pos.payment.terminal.config.form</field>
-            <field name="model">pos.config</field>
-            <field name="inherit_id" ref="point_of_sale.view_pos_config_form"/>
-            <field name="arch" type="xml">
-                <field name="iface_cashdrawer" position="after">
-                    <field name="iface_payment_terminal"/>
-                </field>
-            </field>
-        </record>
-
-        <record id="view_account_journal_pos_user_form" model="ir.ui.view">
-            <field name="name">pos.payment.terminal.journal.form</field>
-            <field name="model">account.journal</field>
-            <field name="inherit_id" ref="point_of_sale.view_account_journal_pos_user_form"/>
-            <field name="arch" type="xml">
-                <field name="amount_authorized_diff" position="after">
-                    <field name="payment_mode"/>
-                </field>
-            </field>
-        </record>
-
-    </data>
-</openerp>
