--- pos_payment_terminal/static/src/xml/templates.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <template id="assets" inherit_id="point_of_sale.assets">
-          <xpath expr="." position="inside">
-              <script type="text/javascript" src="/pos_payment_terminal/static/src/js/pos_payment_terminal.js"></script>
-              <link rel="stylesheet" href="/pos_payment_terminal/static/src/css/pos_payment_terminal.css" />
-          </xpath>
-        </template>
-
-    </data>
-</openerp>
