+++ pos_payment_terminal/views/pos_config.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="view_pos_config_form" model="ir.ui.view">
+        <field name="name">pos.payment.terminal.config.form</field>
+        <field name="model">pos.config</field>
+        <field name="inherit_id" ref="point_of_sale.view_pos_config_form"/>
+        <field name="arch" type="xml">
+            <field name="iface_cashdrawer" position="after">
+                <field name="iface_payment_terminal"/>
+            </field>
+        </field>
+    </record>
+
+</odoo>
